+++ integration_tests/test_model.py
@@ -4,12 +4,12 @@
 # SPDX-License-Identifier: Apache-2.0
 import pytest
 
-from datacube.model import Dataset, DatasetType
+from datacube.model import Dataset, Product
 from typing import List
 
 
 @pytest.mark.parametrize('datacube_env_name', ('datacube', ))
-def test_crs_parse(indexed_ls5_scene_products: List[DatasetType]) -> None:
+def test_crs_parse(indexed_ls5_scene_products: List[Product]) -> None:
     product = indexed_ls5_scene_products[2]
 
     # Explicit CRS, should load fine.